requirements.txt
pipecat-ai[webrtc,daily,silero,deepgram,openai,cartesia,local-smart-turn-v3,runner,google,koala,livekit,sarvam,aic]>= 0.0.97
aiofiles>=23.2.0
python-dotenv>=1.0.0
loguru>=0.7.2
httpx>=0.25.0
fastapi>=0.104.0
aiohttp>=3.9.0
uvicorn[standard]>=0.24.0
boto3

